Hotfix for sending results to sauce labs
The is local check in SauceLabsListener was the opposite of what it
should be. This caused a print of a NullPointerException for local test
runs and not sending results to sauce labs for remote runs.
This hotfix fixes both of those issues and adds a convenience isLocal()
method to ConductorConfig (with unit tests).