start session retries
Running tests remotely can sometimes cause the sessions to not be
created correctly. I suspect this is due to network problems.

This feature allows to retry starting the session if it fails.
The maximum amount of retries can be set in the config with the
value 'startSessionRetries' and defaults to '1'.
which specifies how often